refactor(theme): refactor palette functions and clean up code
